7.15. Отрисуй меня полностью
